+++ src/Pim/Bundle/CatalogBundle/Entity/GroupTranslation.php
@@ -33,7 +33,7 @@ class GroupTranslation extends AbstractTranslation
      * Change foreign key to add constraint and work with basic entity
      *
      * @ORM\ManyToOne(targetEntity="Group", inversedBy="translations")
-     * @ORM\JoinColumn(name="foreign_key", referencedColumnName="id")
+     * @ORM\JoinColumn(name="foreign_key", referencedColumnName="id", onDelete="CASCADE")
      */
     protected $foreignKey;