+++ src/Pim/Bundle/ImportExportBundle/Archiver/JobExecutionArchiver.php
@@ -5,6 +5,7 @@ namespace Pim\Bundle\ImportExportBundle\Archiver;
 use Oro\Bundle\BatchBundle\Entity\JobExecution;
 use Pim\Bundle\ImportExportBundle\Reader\File\CsvReader;
 use Pim\Bundle\ImportExportBundle\Writer\File\FileWriter;
+use Pim\Bundle\ImportExportBundle\Writer\File\ArchivableWriterInterface;
 
 /**
  * Archive job execution files into conventional directories
@@ -50,26 +51,14 @@ class JobExecutionArchiver
             }
             if ($writer instanceof FileWriter) {
                 $sourcePath = $writer->getPath();
-                if (file_exists($sourcePath)) {
+                if ($writer instanceof ArchivableWriterInterface && count($writer->getWrittenFiles()) > 1) {
+                    $archivePath = sprintf('%s/%s.zip', $archivePath, pathinfo($sourcePath, PATHINFO_FILENAME));
+                    $this->createZipArchive($writer->getWrittenFiles(), $archivePath);
+                } elseif (file_exists($sourcePath)) {
                     $this->copyFile($sourcePath, $archivePath);
                 }
             }
         }
-    }
-
-    /**
-     * Copy the source path to the archive
-     * @param string $sourcePath
-     * @param string $archivePath
-     */
-    protected function copyFile($sourcePath, $archivePath)
-    {
-        $sourceName = basename($sourcePath);
-        $destPath   = $archivePath.$sourceName;
-        if (!is_dir($archivePath)) {
-            mkdir($archivePath, 0777, true);
-        }
-        copy($sourcePath, $destPath);
     }
 
     /**
@@ -114,5 +103,62 @@ class JobExecutionArchiver
     public function getBaseDirectory($jobType)
     {
         return $this->rootDir.DIRECTORY_SEPARATOR.$jobType.DIRECTORY_SEPARATOR;
+    }
+
+    /**
+     * Copy the source path to the archive
+     * @param string $sourcePath
+     * @param string $archivePath
+     */
+    protected function copyFile($sourcePath, $archivePath)
+    {
+        $this->ensureDir($archivePath);
+        $sourceName = basename($sourcePath);
+        $destPath   = $archivePath.$sourceName;
+        copy($sourcePath, $destPath);
+    }
+
+    /**
+     * Ensure that a directory exists - if it doesn't, it will be created
+     * @param string $directory
+     */
+    protected function ensureDir($directory)
+    {
+        if (!is_dir($directory)) {
+            mkdir($directory, 0755, true);
+        }
+    }
+
+    /**
+     * Create a zip archive with the execution results.
+     *
+     * @param array  $writtenFiles
+     * @param string $archivePath
+     *
+     * @throws \RuntimeException If an error occurs when creating the archive
+     */
+    protected function createZipArchive($writtenFiles, $archivePath)
+    {
+        $this->ensureDir(pathinfo($archivePath, PATHINFO_DIRNAME));
+
+        $archive = new \ZipArchive();
+        $status = $archive->open($archivePath, \ZIPARCHIVE::CREATE);
+        if ($status !== true) {
+            throw new \RuntimeException(sprintf('Error "%d" occured when creating the zip archive.', $status));
+        }
+
+        foreach ($writtenFiles as $fullPath => $localPath) {
+            $status = $archive->addFile($fullPath, $localPath);
+            if ($status !== true) {
+                throw new \RuntimeException(
+                    sprintf(
+                        'Unknown error occured when adding file "%s" to the zip archive.',
+                        $fullPath
+                    )
+                );
+            }
+        }
+
+        $archive->close();
     }
 }