+++ src/Pim/Bundle/ImportExportBundle/Normalizer/AttributeNormalizer.php
@@ -27,11 +27,6 @@ class AttributeNormalizer implements NormalizerInterface
     protected $supportedFormats = array('json', 'xml');
 
     /**
-     * @var array
-     */
-    protected $results;
-
-    /**
      * Transforms an object into a flat array
      *
      * @param ProductAttribute $attribute
@@ -42,46 +37,34 @@ class AttributeNormalizer implements NormalizerInterface
      */
     public function normalize($attribute, $format = null, array $context = array())
     {
-        $attributeTypes = explode('_', $attribute->getAttributeType());
-
-        $dateMin = (is_null($attribute->getDateMin())) ? '' : $attribute->getDateMin()->format(\DateTime::ISO8601);
-        $dateMax = (is_null($attribute->getDateMax())) ? '' : $attribute->getDateMax()->format(\DateTime::ISO8601);
-
-        $this->results = array(
-            'type'                    => end($attributeTypes),
-            'code'                    => $attribute->getCode(),
-            'label'                   => $this->normalizeLabel($attribute),
-            'available_locales'       => $this->normalizeAvailableLocales($attribute),
-            'group'                   => $attribute->getVirtualGroup()->getCode(),
-            'sort_order'              => $attribute->getSortOrder(),
-            'required'                => $attribute->getRequired(),
-            'unique'                  => $attribute->getUnique(),
-            'searchable'              => $attribute->getSearchable(),
-            'localizable'             => $attribute->getTranslatable(),
-            'scope'                   => $attribute->getScopable() ? self::CHANNEL_SCOPE : self::GLOBAL_SCOPE,
-            'useable_as_grid_column'  => (string) (int) $attribute->isUseableAsGridColumn(),
-            'useable_as_grid_filter'  => (string) (int) $attribute->isUseableAsGridFilter(),
-            'default_value'           => (string) $attribute->getDefaultValue(),
-            'max_characters'          => (string) $attribute->getMaxCharacters(),
-            'validation_rule'         => (string) $attribute->getValidationRule(),
-            'validation_regexp'       => (string) $attribute->getValidationRegexp(),
-            'wysiwyg_enabled'         => (string) $attribute->isWysiwygEnabled(),
-            'number_min'              => (string) $attribute->getNumberMin(),
-            'number_max'              => (string) $attribute->getNumberMax(),
-            'decimals_allowed'        => (string) $attribute->isDecimalsAllowed(),
-            'negative_allowed'        => (string) $attribute->isNegativeAllowed(),
-            'date_min'                => $dateMin,
-            'date_max'                => $dateMax,
-            'date_type'               => (string) $attribute->getDateType(),
-            'metric_family'           => (string) $attribute->getMetricFamily(),
-            'default_metric_unit'     => (string) $attribute->getDefaultMetricUnit(),
-            'allowed_extensions'      => implode(self::ITEM_SEPARATOR, $attribute->getAllowedExtensions()),
-            'max_file_size'           => (string) $attribute->getMaxFileSize(),
-            'options'                 => $this->normalizeOptions($attribute),
-            'default_options'         => $this->normalizeDefaultOptions($attribute)
+        $results = array(
+            'type' => $attribute->getAttributeType(),
+            'code' => $attribute->getCode()
         );
+        $results = array_merge($results, $this->normalizeLabel($attribute));
+        $results = array_merge(
+            $results,
+            array(
+                'group'                   => $attribute->getVirtualGroup()->getCode(),
+                'unique'                  => (int) $attribute->getUnique(),
+                'useable_as_grid_column'  => (int) $attribute->isUseableAsGridColumn(),
+                'useable_as_grid_filter'  => (int) $attribute->isUseableAsGridFilter(),
+            )
+        );
+        if (isset($context['versioning'])) {
+            $results = array_merge($results, $this->getVersionedData($attribute));
 
-        return $this->results;
+        } else {
+            $results = array_merge(
+                $results,
+                array(
+                    'is_translatable' => (int) $attribute->getTranslatable(),
+                    'is_scopable'     => (int) $attribute->getScopable(),
+                )
+            );
+        }
+
+        return $results;
     }
 
     /**
@@ -98,6 +81,46 @@ class AttributeNormalizer implements NormalizerInterface
     }
 
     /**
+     * Get extra data to store in version
+     *
+     * @param ProductAttribute $attribute
+     *
+     * @return array
+     */
+    protected function getVersionedData(ProductAttribute $attribute)
+    {
+        $dateMin = (is_null($attribute->getDateMin())) ? '' : $attribute->getDateMin()->format(\DateTime::ISO8601);
+        $dateMax = (is_null($attribute->getDateMax())) ? '' : $attribute->getDateMax()->format(\DateTime::ISO8601);
+
+        return array(
+            'available_locales'   => $this->normalizeAvailableLocales($attribute),
+            'searchable'          => $attribute->getSearchable(),
+            'localizable'         => $attribute->getTranslatable(),
+            'scope'               => $attribute->getScopable() ? self::CHANNEL_SCOPE : self::GLOBAL_SCOPE,
+            'options'             => $this->normalizeOptions($attribute),
+            'default_options'     => $this->normalizeDefaultOptions($attribute),
+            'sort_order'          => (int) $attribute->getSortOrder(),
+            'required'            => (int) $attribute->getRequired(),
+            'default_value'       => (string) $attribute->getDefaultValue(),
+            'max_characters'      => (string) $attribute->getMaxCharacters(),
+            'validation_rule'     => (string) $attribute->getValidationRule(),
+            'validation_regexp'   => (string) $attribute->getValidationRegexp(),
+            'wysiwyg_enabled'     => (string) $attribute->isWysiwygEnabled(),
+            'number_min'          => (string) $attribute->getNumberMin(),
+            'number_max'          => (string) $attribute->getNumberMax(),
+            'decimals_allowed'    => (string) $attribute->isDecimalsAllowed(),
+            'negative_allowed'    => (string) $attribute->isNegativeAllowed(),
+            'date_min'            => $dateMin,
+            'date_max'            => $dateMax,
+            'date_type'           => (string) $attribute->getDateType(),
+            'metric_family'       => (string) $attribute->getMetricFamily(),
+            'default_metric_unit' => (string) $attribute->getDefaultMetricUnit(),
+            'allowed_extensions'  => implode(self::ITEM_SEPARATOR, $attribute->getAllowedExtensions()),
+            'max_file_size'       => (string) $attribute->getMaxFileSize(),
+        );
+    }
+
+    /**
      * Normalize the label
      *
      * @param ProductAttribute $attribute
@@ -111,7 +134,7 @@ class AttributeNormalizer implements NormalizerInterface
             $labels[$translation->getLocale()]= $translation->getLabel();
         }
 
-        return $labels;
+        return array('label' => $labels);
     }
 
     /**