+++ src/Pim/Bundle/ImportExportBundle/Reader/ORM/EntityReader.php
@@ -11,7 +11,7 @@ use Doctrine\ORM\EntityManager;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class EntityReader extends CursorReader
+class EntityReader extends Reader
 {
     /**
      * @var EntityManager