+++ src/Pim/Bundle/ImportExportBundle/Tests/Unit/Normalizer/AttributeNormalizerTest.php
@@ -50,7 +50,7 @@ class AttributeNormalizerTest extends NormalizerTestCase
         return array(
             array(
                 array(
-                    'type'                   => 'multiselect',
+                    'type'                   => 'pim_catalog_multiselect',
                     'code'                   => 'color',
                     'label'                  => array('en' => 'Color', 'fr' => 'Couleur'),
                     'group'                  => 'general',
@@ -72,7 +72,7 @@ class AttributeNormalizerTest extends NormalizerTestCase
             ),
             array(
                 array(
-                    'type'                   => 'text',
+                    'type'                   => 'pim_catalog_text',
                     'code'                   => 'description',
                     'label'                  => array('en' => 'Description', 'fr' => 'Description'),
                     'group'                  => 'info',
@@ -113,7 +113,7 @@ class AttributeNormalizerTest extends NormalizerTestCase
             }
         }
 
-        $normalized = $this->normalizer->normalize($attribute, $this->format);
+        $normalized = $this->normalizer->normalize($attribute, $this->format, array('versioning' => true));
         $this->assertEquals($expectedResult, $normalized);
     }
 
@@ -151,7 +151,7 @@ class AttributeNormalizerTest extends NormalizerTestCase
     protected function createEntity(array $data)
     {
         $attribute = new ProductAttribute();
-        $attribute->setAttributeType(sprintf('pim_catalog_%s', strtolower($data['type'])));
+        $attribute->setAttributeType($data['type']);
 
         $this->addLabels($attribute, $data);